Backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pypdf2